requirements.txt
Pillow==9.2.0
PyPDF2==2.8.1
pytube==12.1.0
pytube3==9.6.4
